docker/requirements.txt
gnupg==2.3.1
python-debian==0.1.49
GitPython==3.1.31
